FlaskStuff/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
